pyproject.toml
[tool.poetry]
name = "scraping-datafootball"
version = "0.1.0"
description = ""
authors = ["Gabriel <gabriel.albuquerque05@gmail.com>"]
readme = "README.md"
packages = [{include = "scraping_datafootball", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
pandas = "^2.2.3"
requests = "^2.32.3"
curl-cffi = "^0.10.0"
brotli = "^1.1.0"
python-dotenv = "^1.1.0"
unidecode = "^1.3.8"
selenium = "^4.31.0"
playwright = "^1.51.0"
cfscrape = "^2.1.1"
tqdm = "^4.67.1"
sqlalchemy = "1.4.49"
apache-airflow = "^3.0.3"
psycopg2-binary = "^2.9.10"
asyncpg = "^0.30.0"
boto3 = "^1.40.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"